requirements.txt
yt-dlp>=2024.1.0
google-api-python-client>=2.100.0
openai>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.7.0
